Started file organization
