+++ axum-webserver/src/error.rs
@@ -1,0 +1,9 @@
+use axum::http::StatusCode;
+
+/// Utility function for mapping any error into a `500 Internal Server Error` response.
+pub fn internal_error<E>(err: E) -> (StatusCode, String)
+where
+    E: std::error::Error,
+{
+    (StatusCode::INTERNAL_SERVER_ERROR, err.to_string())
+}
