+++ axum-webserver/src/main.rs
@@ -1,32 +1,21 @@
 use axum::{
     Router,
-    extract::State,
-    http::StatusCode,
-    response::Json,
     routing::{get, post},
 };
-use diesel::prelude::*;
 use diesel_migrations::{EmbeddedMigrations, embed_migrations};
 use std::net::SocketAddr;
 
-// Import the schema module generated by Diesel
+mod db;
+mod error;
+mod models;
+mod routes;
 mod schema;
-use schema::users;
 
-// Define migrations dir for diesel
+use routes::health::health_check;
+use routes::user::{create_user, list_users};
+
+// Diesel embedded migrations
 pub const MIGRATIONS: EmbeddedMigrations = embed_migrations!("migrations/");
-
-#[derive(serde::Serialize, Selectable, Queryable)]
-struct User {
-    id: i32,
-    username: String,
-}
-
-#[derive(serde::Deserialize, Insertable)]
-#[diesel(table_name = users)]
-struct NewUser {
-    username: String,
-}
 
 #[tokio::main]
 async fn main() {
@@ -48,46 +37,4 @@ async fn main() {
     let listener = tokio::net::TcpListener::bind(addr).await.unwrap();
     println!("listening on {}", listener.local_addr().unwrap());
     axum::serve(listener, app).await.unwrap();
-}
-
-async fn health_check() -> (StatusCode, &'static str) {
-    (StatusCode::OK, "OK")
-}
-
-async fn create_user(
-    State(pool): State<deadpool_diesel::postgres::Pool>,
-    Json(new_user): Json<NewUser>,
-) -> Result<Json<User>, (StatusCode, String)> {
-    let conn = pool.get().await.map_err(internal_error)?;
-    let res = conn
-        .interact(|conn| {
-            diesel::insert_into(users::table)
-                .values(new_user)
-                .returning(User::as_returning())
-                .get_result(conn)
-        })
-        .await
-        .map_err(internal_error)? // First map handles errors with connection pool
-        .map_err(internal_error)?; // Second map handles errors with interact
-    Ok(Json(res))
-}
-
-async fn list_users(
-    State(pool): State<deadpool_diesel::postgres::Pool>,
-) -> Result<Json<Vec<User>>, (StatusCode, String)> {
-    let conn = pool.get().await.map_err(internal_error)?;
-    let res = conn
-        .interact(|conn| users::table.select(User::as_select()).load(conn))
-        .await
-        .map_err(internal_error)?
-        .map_err(internal_error)?;
-    Ok(Json(res))
-}
-
-/// Utility function for mapping any error into a `500 Internal Server Error` response.
-fn internal_error<E>(err: E) -> (StatusCode, String)
-where
-    E: std::error::Error,
-{
-    (StatusCode::INTERNAL_SERVER_ERROR, err.to_string())
 }
