+++ axum-webserver/src/models/mod.rs
@@ -1,2 +1,1 @@
-pub mod todo;
 pub mod user;
