+++ axum-webserver/src/routes/health.rs
@@ -1,0 +1,5 @@
+use axum::http::StatusCode;
+
+pub async fn health_check() -> (StatusCode, &'static str) {
+    (StatusCode::OK, "OK")
+}
